requirements.txt
requests
bs4
flask
flask-sqlalchemy
flask-login
mysql-connector-python
apscheduler
python-dotenv